backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# PDF Generation
reportlab>=4.0.0
weasyprint>=60.0

# Task Queue (for delayed alerts)
celery>=5.3.0
redis>=5.0.0

# Utilities
python-decouple>=3.8
pillow>=10.0.0
django-filter>=23.5

# Development
django-extensions>=3.2.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
